# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/writing-stats/upsert": {
      "post": {
        "description": "\n            Une seule entrée par utilisateur + livre + jour.\n            Si une entrée existe déjà pour ce jour, les mots et minutes sont additionnés.\n            Date par défaut : aujourd'hui.\n    ",
        "operationId": "WritingStatController_upsert",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertWritingStatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session enregistrée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enregistrer une session d'écriture (upsert par jour)",
        "tags": [
          "Writing Stats"
        ]
      }
    },
    "/api/v1/writing-stats/summary": {
      "get": {
        "description": "\n            Retourne :\n            - Total mots écrits, minutes passées, jours actifs\n            - Moyenne mots/jour et minutes/jour\n            - Meilleur jour (plus de mots écrits)\n            - Streak actuel (jours consécutifs d'écriture)\n    ",
        "operationId": "WritingStatController_getSummary",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Résumé des statistiques d'écriture",
        "tags": [
          "Writing Stats"
        ]
      }
    },
    "/api/v1/writing-stats/daily": {
      "get": {
        "description": "Retourne une entrée par jour avec mots écrits et minutes passées.",
        "operationId": "WritingStatController_getDaily",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Statistiques jour par jour (pour graphique barres / courbe)",
        "tags": [
          "Writing Stats"
        ]
      }
    },
    "/api/v1/writing-stats/by-book": {
      "get": {
        "description": "\n            Retourne pour chaque livre :\n            - Total mots écrits et heures passées\n            - Nombre de jours actifs\n            - Moyenne mots par jour\n    ",
        "operationId": "WritingStatController_getByBook",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Statistiques agrégées par livre",
        "tags": [
          "Writing Stats"
        ]
      }
    },
    "/api/v1/writing-stats/streak": {
      "get": {
        "description": "Compte les jours consécutifs où l'utilisateur a écrit au moins 1 mot.",
        "operationId": "WritingStatController_getStreak",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookId",
            "required": false,
            "in": "query",
            "description": "Streak spécifique à un livre",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Calculer le streak d'écriture actuel (jours consécutifs)",
        "tags": [
          "Writing Stats"
        ]
      }
    },
    "/api/v1/writing-stats/{statId}": {
      "delete": {
        "operationId": "WritingStatController_remove",
        "parameters": [
          {
            "name": "statId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer une entrée de statistique",
        "tags": [
          "Writing Stats"
        ]
      }
    },
...
  },
  "info": {
    "title": "projet-x-service",
    "description": "API du projet-x-service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpsertWritingStatDto": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "string",
            "description": "ID du livre concerné (optionnel — session globale si absent)",
            "example": "clxBOOK001"
          },
          "date": {
            "type": "string",
            "description": "Date de la session (YYYY-MM-DD) — aujourd'hui par défaut",
            "example": "2025-04-12"
          },
          "wordsWritten": {
            "type": "number",
            "description": "Nombre de mots écrits durant la session",
            "example": 350,
            "minimum": 0
          },
          "minutesSpent": {
            "type": "number",
            "description": "Temps passé à écrire en minutes",
            "example": 45,
            "minimum": 0
          }
        },
        "required": [
          "wordsWritten",
          "minutesSpent"
        ]
      },
...
    }
  }
}